Revert "Add mistral (#38)"
This reverts commit 943896fa006d1adcb0de5cf8c03d077958034d9e.
