+++ src/app/globals.css
@@ -40,7 +40,7 @@
 
 @layer base {
   :root {
-    --background: 0 35% 99%;
+    --background: 0 0% 93%;
     --foreground: 240 10% 3.9%;
     --card: 0 0% 100%;
     --card-foreground: 240 10% 3.9%;
