+++ src/features/chat-page/chat-services/models.ts
@@ -90,12 +90,6 @@ export type AzureChatCompletionContent = {
   response: ChatCompletionSnapshot;
 };
 
-export type AzureChatLLMCompletionContent = {
-  type: "contentLLM";
-  response: string;
-  idMessage: string;
-};
-
 export type AzureChatCompletionFinalContent = {
   type: "finalContent";
   response: string;
@@ -117,5 +111,4 @@ export type AzureChatCompletion =
   | AzureChatCompletionFunctionCallResult
   | AzureChatCompletionContent
   | AzureChatCompletionFinalContent
-  | AzureChatCompletionAbort
-  | AzureChatLLMCompletionContent;
+  | AzureChatCompletionAbort;
