+++ src/features/chat-page/tokenCount/utils.ts
@@ -1,6 +1,6 @@
-import { GPT } from "@/features/theme/theme-config";
 import { ChatMessageModel } from "../chat-services/models";
 import { getEncoding } from 'js-tiktoken';
+import { GPT } from '@/features/common/services/openai';
 const encoder = getEncoding('cl100k_base');
 
 // https://github.com/dqbd/tiktoken/issues/23#issuecomment-1483317174
