+++ src/package.json
@@ -10,10 +10,7 @@
     "debug": "set NODE_OPTIONS=--inspect && next dev"
   },
   "dependencies": {
-    "@azure-rest/ai-inference": "^1.0.0-beta.5",
     "@azure/ai-form-recognizer": "^5.0.0",
-    "@azure/core-auth": "^1.9.0",
-    "@azure/core-sse": "^2.1.3",
     "@azure/cosmos": "^4.0.0",
     "@azure/identity": "^4.4.1",
     "@azure/keyvault-secrets": "^4.7.0",
@@ -49,7 +46,7 @@
     "next": "^14.2.24",
     "next-auth": "^4.24.5",
     "next-themes": "^0.2.1",
-    "openai": "^4.67.1",
+    "openai": "^4.80.0",
     "react": "^18",
     "react-dom": "^18",
     "react-syntax-highlighter": "^15.5.0",
