+++ src/types/type.ts
@@ -5,11 +5,6 @@ const azureEnvVars = [
   "AZURE_OPENAI_API_DEPLOYMENT_NAME_4omini",
   "AZURE_OPENAI_API_DEPLOYMENT_NAME_4o",
   "AZURE_OPENAI_API_DEPLOYMENT_NAME_o1mini",
-  "AZURE_LLM_API_INSTANCE_NAME",
-  "AZURE_LLM_API_ENDPOINT_SUFFIX",
-  "AZURE_AI_API_KEY",
-  "AZURE_AI_API_DEPLOYMENT_NAME_Codestral",
-  "AZURE_AI_API_DEPLOYMENT_NAME_MistralNemo",
   "AZURE_OPENAI_API_VERSION",
   "AZURE_OPENAI_API_EMBEDDINGS_DEPLOYMENT_NAME",
   "AZURE_COSMOSDB_URI",
